requirements.txt
boto3 == 1.36.4
PyPDF2 == 3.0.1
python-dotenv~=1.0.1
Flask~=3.1.0
langid~=1.1.6